+++ NEWS.md
@@ -2,6 +2,7 @@
 *   More detailed work-throughs of examples in the vignettes
 *   Fixed typo in Bird_Parasite_CRF vignette
 *   Fixed typo in cv_MRF_diag function
+*   Changed the root mean square transformation for Poisson variables to a more appropriate paranormal
 
 # MRFcov 1.0.36
 *   Improved clarity of function definitions and descriptions